bug fixed

+++ main.py
@@ -17,7 +17,8 @@ def main_menu():
     print('    3. Quit CatCoin')
     choice = input()
     while choice != '1' and choice != '2' and choice != '3':
-        choice = input(print('Invalid input. Try again: '))
+        print('Invalid input. Try again: ')
+        choice = input()
     if choice == '1':
         check_balance()
     elif choice == '2':
@@ -35,7 +36,8 @@ def check_balance():
     print('Which account would you like to check? (1/2/3)')
     account_number = input()
     while account_number != '1' and account_number != '2' and account_number != '3':
-        account_number = input(print('Invalid input. Try again: '))
+        print('Invalid input. Try again: ')
+        account_number = input()
     with open('balance.csv', 'r') as f:
         balance_line = f.readlines()
         for line in balance_line:
@@ -52,7 +54,8 @@ def check_balance():
     print('    2. Return to Main Menu')
     check_other = input()
     while check_other != '1' and check_other != '2':
-        check_other = input(print('Invalid input. Try again: '))
+        print('Invalid input. Try again: ')
+        check_other = input()
     if check_other == '1':
         check_balance()
     else:
@@ -69,41 +72,49 @@ def transfer():
             wallet = bal[0]
             balance = bal[1]
             balance = "".join(balance).replace('\n', '')
-            balance_dict[wallet] = int(balance)
+            balance_dict[wallet] = float(balance)
     print('Choose a source wallet 1/2/3: ')
     source = input()
     while source != '1' and source != '2' and source != '3':
-        source = input(print('Invalid input. Try again: '))
+        print('Invalid input. Try again: ')
+        source = input()
     print('Choose a destination wallet 1/2/3: ')
     destination = input()
     while destination != '1' and destination != '2' and destination != '3' and destination == source:
-        destination = input(print('Invalid input. Try again: '))
+        print('Invalid input. Try again: ')
+        destination = input()
     try:
         print('Transfer amount: ')
         amount = float(input())
     except ValueError:
-        amount = input(print('Invalid input. Try again: '))
-    print('Are you sure you want to transfer', amount, 'CatCoins to Wallet', destination, '? y/n')
-    decision = input()
-    if decision == 'y' or decision == 'Y':
-        if amount > balance_dict[source]:
-            print('Insufficient balance.')
-            print('-----------------------------------------')
-            main_menu()
-        else:
-            balance_dict[source] -= amount
-            balance_dict[destination] += amount
-            header = ['wallet number', 'balance']
-            data = [{'wallet number': '1', 'balance': balance_dict['1']},
-                     {'wallet number': '2', 'balance': balance_dict['2']},
-                     {'wallet number': '3', 'balance': balance_dict['3']}]
-            with open('balance.csv', 'w') as f:
-                writer = csv.DictWriter(f, fieldnames=header)
-                writer.writeheader()
-                writer.writerows(data)
-    print('Transaction complete')
-    print('-----------------------------------------')
-    main_menu()
+        print('Invalid input. Try again: ')
+        amount = float(input())
+    if amount <= 0:
+        print('Insufficient balance.')
+        print('-----------------------------------------')
+        main_menu()
+    else:
+        print('Are you sure you want to transfer', amount, 'CatCoins to Wallet', destination, '? y/n')
+        decision = input()
+        if decision == 'y' or decision == 'Y':
+            if amount > balance_dict[source]:
+                print('Insufficient balance.')
+                print('-----------------------------------------')
+                main_menu()
+            else:
+                balance_dict[source] -= amount
+                balance_dict[destination] += amount
+                header = ['wallet number', 'balance']
+                data = [{'wallet number': '1', 'balance': balance_dict['1']},
+                        {'wallet number': '2', 'balance': balance_dict['2']},
+                        {'wallet number': '3', 'balance': balance_dict['3']}]
+                with open('balance.csv', 'w') as f:
+                    writer = csv.DictWriter(f, fieldnames=header)
+                    writer.writeheader()
+                    writer.writerows(data)
+        print('Transaction complete')
+        print('-----------------------------------------')
+        main_menu()
 
 
 def main():
